+++ crates/openapi/src/openapi_v2.rs
@@ -178,6 +178,7 @@ Never share your secret api keys. Keep them guarded and secure.
 
         // Routes for Revenue Recovery flow under Process Tracker
         routes::revenue_recovery::revenue_recovery_pt_retrieve_api,
+        routes::revenue_recovery::revenue_recovery_payments_create,
 
         // Routes for proxy
         routes::proxy::proxy_core,
@@ -378,6 +379,7 @@ Never share your secret api keys. Keep them guarded and secure.
         api_models::enums::PaymentMethod,
         api_models::enums::PaymentMethodIssuerCode,
         api_models::enums::MandateStatus,
+        api_models::enums::InvoiceStatus,
         api_models::enums::MerchantProductType,
         api_models::enums::PaymentExperience,
         api_models::enums::CustomerAcceptanceSupport,
@@ -984,6 +986,11 @@ Never share your secret api keys. Keep them guarded and secure.
         api_models::payment_methods::PaymentMethodSessionResponse,
         api_models::payment_methods::AuthenticationDetails,
         api_models::process_tracker::revenue_recovery::RevenueRecoveryResponse,
+        api_models::payments::RecoveryPaymentsCreate,
+        api_models::payments::RecoveryPaymentsResponse,
+        api_models::payments::CustomRecoveryPaymentMethodData,
+        api_models::payments::AdditionalCardInfo,
+        common_types::payments::RecoveryAction,
         api_models::enums::RevenueRecoveryAlgorithmType,
         api_models::enums::ProcessTrackerStatus,
         api_models::proxy::ProxyRequest,